Created List of Usernames
This uses the first names and surnames file to generate a list of
usernames which would correspond to the first names and surnames in
sequence i.e if there are 1000 first names and 1100 surnames then the
script will loop the first names so the last 100 surnames will be
paired with the first 100 first names. This loop continues depending on
how many usernames the developer specifies in the iterator.

The usernames are in the form
First name: Joe
Last name: Bloggs
Username: JoeBloggs