Built the data generator
Five modes of operation
 - RANDOMRANGE
 - RANDOMFLOATRANGE
 - RANDOMSEQ
 - RANDOMCHAR
 - RANDOMBOOL

These are used for generating randomised data for fields which do not
require an isolated file. RANDOMRANGE and RANDOMFLOATRANGE generates
random int or float values within the specified range. RANDOMSEQ
generates sequential list of integers with each calling of the
function. Once the sequence reaches the maximum value the first value
is returned and the sequence starts again. RANDOMBOOL returns true or
false values